+++ src/effects/reverse.rs
@@ -40,7 +40,7 @@ impl EffectTrait for Reverse {
         Err("Reverse has no configurable parameters".to_string())
     }
 
-    fn apply(&self, samples: &mut Vec<f64>, _sample_rate: u32) -> Result<(), &'static str> {
+    fn apply(&self, samples: &mut Vec<f32>, _sample_rate: u32) -> Result<(), &'static str> {
         samples.reverse();
         Ok(())
     }
