+++ src/effects/tremolo.rs
@@ -40,13 +40,13 @@ impl EffectTrait for Tremolo {
         Err("Tremolo has no configurable parameters".to_string())
     }
 
-    fn apply(&self, samples: &mut Vec<f64>, sample_rate: u32) -> Result<(), &'static str> {
-        const FREQUENCY: f64 = 8.0; // Tremolo frequency in Hz
-        const DEPTH: f64 = 0.3; // Tremolo depth (0.0 to 1.0)
+    fn apply(&self, samples: &mut Vec<f32>, sample_rate: u32) -> Result<(), &'static str> {
+        const FREQUENCY: f32 = 8.0; // Tremolo frequency in Hz
+        const DEPTH: f32 = 0.3; // Tremolo depth (0.0 to 1.0)
 
         for (i, sample) in samples.iter_mut().enumerate() {
-            let time = i as f64 / sample_rate as f64;
-            let angle = 2.0 * std::f64::consts::PI * FREQUENCY * time;
+            let time = i as f32 / sample_rate as f32;
+            let angle = 2.0 * std::f32::consts::PI * FREQUENCY * time;
             let mod_factor = 1.0 - (DEPTH * (0.5 + 0.5 * angle.sin()));
             *sample *= mod_factor;
         }
